requirements.txt
pandas==2.0.2
python-decouple==3.8